refactor : template, requestform 분리
