+++ index.html
@@ -53,6 +53,6 @@
         </div>
       </section>
     </div>
-    <script src="./src/js/index.js"></script>
+    <script type="module" src="./src/js/index.js"></script>
   </body>
 </html>
